Cached and disk zone files renamed
